Txt / Csv parsers fixes

+++ manager/uploads/parsers/csv.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import numpy as np
 from overrides import overrides
 from manager.uploads.parsers.txt import Txt
 
@@ -10,4 +11,4 @@ class Csv(Txt):
 
     @overrides
     def readPandas(self, fileForPandas, skipRows):
-        return pd.read_csv(fileForPandas, sep=',', header=None, skiprows=skipRows)
+        return pd.read_csv(fileForPandas, sep=',', header=None, skiprows=skipRows, dtype=np.float64)
